refactoring of classes
